refactor: rename for clarity
forward_incoming_proposal -> relay_incoming_proposal: the function
opens a consensus session and emits a UI banner, so "forward"
misleads.
auto_approved_leave_proposal_id -> self_leave_proposal_id: drops
implementation detail from the public name; matches the existing
is_pending_self_leave naming.
app/peer_scoring.rs -> app/peer_scoring_backends.rs: removes the
parallel name with core/peer_scoring.rs and reflects the file's
actual scope (storage + delta-provider impls, not the trait).
